% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{address}
\alias{address}
\title{Sample Data: Addresses}
\format{An object of class \code{"character"}.}
\source{
\code{save(address, file="data/address.RData")}
}
\usage{
data(address)
}
\description{
Some addresses inside the boundaries of the Zurich city, Switzerland.
}
\examples{
data(address)
}
\keyword{datasets}
