% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defunct.R
\name{autocomplete-defunct}
\alias{autocomplete-defunct}
\title{HERE Geocoder API: Autocomplete}
\usage{
autocomplete(addresses, results, url_only)
}
\arguments{
\item{addresses}{character, addresses to autocomplete.}

\item{results}{numeric, maximum number of suggestions (Valid range: 1 and 20).}

\item{url_only}{boolean, only return the generated URLs (\code{default = FALSE})?}
}
\value{
A \code{data.frame} object, containing the autocomplete suggestions for the addresses.
}
\description{
Completes addresses using the HERE 'Geocoder Autocomplete' API.
}
\references{
\href{https://developer.here.com/documentation/geocoder-autocomplete/dev_guide/topics/resource-suggest.html}{HERE Geocoder API: Autocomplete}
}
\seealso{
\code{\link{hereR-defunct}}
}
\keyword{internal}
