% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_addresses}
\alias{example_addresses}
\title{Example Data Set: Addresses}
\format{An object of class \code{"character"}.}
\usage{
data(example_addresses)
}
\description{
Some addresses inside the boundaries of the Zurich city, Switzerland.
}
\examples{
data(example_addresses)
}
\keyword{datasets}
