% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_geocode}
\alias{example_geocode}
\title{Example Data Set: Geocoded POIs}
\format{An object of class \code{"character"}.}
\source{
\code{example_geocode <- geocode(example_addresses)}
}
\usage{
data(example_geocode)
}
\description{
Geocoded Points of Interest (POIs) inside the boundaries of the Zurich city,
Switzerland.
}
\examples{
data(example_geocode)
}
\keyword{datasets}
