% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geocode.R
\name{geocode}
\alias{geocode}
\title{HERE Geocoder API: Geocode}
\usage{
geocode(addresses, url_only = FALSE)
}
\arguments{
\item{addresses}{character, addresses to geocode.}
}
\value{
An sf object, containing the coordinates of the geocoded addresses.
}
\description{
Geocodes addresses using the HERE Geocoder API.
}
\references{
\href{https://developer.here.com/documentation/geocoder/topics/resource-geocode.html}{HERE Geocoder API: Geocode}
}
