% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defunct.R
\name{hereR-defunct}
\alias{hereR-defunct}
\alias{traffic}
\alias{set_auth}
\alias{unset_auth}
\alias{autocomplete}
\alias{set_proxy}
\title{Defunct functions in package \pkg{hereR}.}
\usage{
traffic(
  aoi,
  product = "flow",
  from = NULL,
  to = NULL,
  min_jam_factor = 0,
  url_only = FALSE
)

set_auth(app_id, app_code)

unset_auth()

autocomplete(addresses, results = 5, url_only = FALSE)

set_proxy(proxy, proxyuserpwd)
}
\description{
The functions listed below are defunct.
  When possible, alternative functions with similar
  functionality are also mentioned. Help pages for defunct functions are
  available at \code{help("-defunct")}.
}
\section{\code{traffic}}{

For \code{traffic(..., product = "flow")}, use \code{\link{flow}} and
for \code{traffic(..., product = "incidents")}, use \code{\link{incident}}.
}

\section{\code{set_auth}}{

For \code{set_auth}, use \code{\link{set_key}}.
}

\section{\code{unset_auth}}{

For \code{unset_auth}, use \code{\link{unset_key}}.
}

\section{\code{autocomplete}}{

For \code{autocomplete}, use \code{\link{autosuggest}}.
}

\section{\code{set_proxy}}{

For \code{set_proxy}, configure a global proxy for R in '~/.Renviron' instead.
}

\keyword{internal}
