% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isoline.R
\name{isoline}
\alias{isoline}
\title{HERE Routing API: Isoline}
\usage{
isoline(poi, range = seq(5, 15, 5) * 60, rangetype = "time",
  type = "fastest", mode = "car", traffic = FALSE,
  departure = NULL, start = TRUE, aggregate = TRUE,
  url_only = FALSE)
}
\arguments{
\item{url_only}{boolean, only return the generated URLs (default = FALSE)?}
}
\description{
Calcuates a isoline (POLYGON) that connects the end points of all routes
leaving from defined centers (POIs) with either a specified length or a
specified travel time.
}
\references{
\href{https://developer.here.com/documentation/routing/topics/resource-calculate-isoline.html}{HERE Routing API: Calculate Isoline}
}
