% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{poi}
\alias{poi}
\title{Sample Data: POIs}
\format{An object of class \code{"character"}.}
\source{
\code{save(poi, file="data/poi.RData")}
}
\usage{
data(poi)
}
\description{
Geocoded Points of Interest (POIs) inside the boundaries of the Zurich city,
Switzerland.
}
\examples{
data(poi)
}
\keyword{datasets}
