% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pt_route.R
\name{pt_route}
\alias{pt_route}
\title{HERE Public Transit API: Route}
\usage{
pt_route(
  start,
  destination,
  results = 3,
  changes = -1,
  time = Sys.time(),
  arrival = FALSE,
  url_only = FALSE
)
}
\arguments{
\item{start}{\code{sf} object, Points of Interest (POIs) of geometry type \code{POINT} for the start locations.}

\item{destination}{\code{sf} object, Points of Interest (POIs) of geometry type \code{POINT} for the destination locations.}

\item{results}{numeric, maximum number of suggested public transit routes (Valid range: 1 and 6).}

\item{changes}{numeric, maximum number of changes allowed per route (Valid range: -1 and 6, \code{default = -1}).}

\item{time}{datetime, timestamp of type \code{POSIXct}, \code{POSIXt} for the departure (or arrival if \code{arrival = TRUE}).}

\item{arrival}{boolean, calculate routes for arrival at the defined time (\code{default = FALSE})?}

\item{url_only}{boolean, only return the generated URLs (\code{default = FALSE})?}
}
\value{
An \code{sf} object containing the requested routes.
}
\description{
Calculates public transit route geometries (\code{LINESTRING}) between given pairs of points using the 'Public Transit' API.
}
\examples{
# Provide an API Key for a HERE project
set_key("<YOUR API KEY>")

# Routing
routes <- pt_route(
  start = poi[3:4, ], destination = poi[5:6, ],
  url_only = TRUE
)
}
\references{
\href{https://developer.here.com/documentation/transit/dev_guide/topics/quick-start-routing-1.html}{HERE Public Transit API: Transit Route}
}
