% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/route.R
\name{route}
\alias{route}
\title{HERE Routing API: Route}
\usage{
route(start, destination, type = "fastest", mode = "car",
  traffic = FALSE, vehicle_type = "diesel,5.5", departure = NULL,
  arrival = NULL)
}
\arguments{
\item{arrival}{}
}
\description{
Calcualtes a route between two points.
}
\references{
\href{https://developer.here.com/documentation/routing/topics/resource-calculate-route.html}{HERE Routing API: Calculate Route}
}
