% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix.R
\name{route_matrix}
\alias{route_matrix}
\title{HERE Routing API: Route Matrix}
\usage{
route_matrix(start, destination = start, type = "fastest",
  mode = "car", traffic = FALSE, searchRange = 10000,
  attribute = c("distance", "traveltime"), departure = NULL,
  url_only = FALSE)
}
\arguments{
\item{url_only}{boolean, only return the generated URLs (default = FALSE)?}
}
\description{
Calcuates a matrix of route summaries.
}
\references{
\href{https://developer.here.com/documentation/routing/topics/resource-calculate-matrix.html}{HERE Routing API: Calculate Matrix}
}
