% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defunct.R
\name{set_auth-defunct}
\alias{set_auth-defunct}
\title{Defunct: Set Application Credentials}
\usage{
set_auth(app_id, app_code)
}
\arguments{
\item{app_id}{character, the APP ID from a HERE project.}

\item{app_code}{character, the APP CODE from a HERE project.}
}
\value{
None.
}
\description{
Provide application credentials (APP ID and APP CODE) for a HERE project
of type 'REST & XYZ HUB API/CLI', that will be used to authenticate in
the requests to the APIs.
}
\details{
No login yet? Get your free login here: \href{https://developer.here.com/}{klick}
}
\seealso{
\code{\link{hereR-defunct}}
}
\keyword{internal}
