% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{set_currency}
\alias{set_currency}
\title{Set the currency for HERE API requests}
\usage{
set_currency(currency = NULL)
}
\arguments{
\item{currency}{character, the currency code compliant to ISO 4217 to use in
the requests (default = \code{NULL}, which defaults to the current system
locale settings).}
}
\value{
None.
}
\description{
If the currency is not set using this function call, the currency defined in
the monetary representations in the current locale is used. If the monetary
formatting category \code{"LC_MONETARY"} of the C locale is not set,
\code{"USD"} is set as default.
}
\examples{
set_currency("CHF")
}
