% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{set_freemium}
\alias{set_freemium}
\title{Set whether plan is freemium or not}
\usage{
set_freemium(ans = TRUE)
}
\arguments{
\item{ans}{boolean, use limits or not (default = \code{TRUE})?}
}
\value{
None.
}
\description{
If set to \code{TRUE} the hereR package limits the requests per second (RPS)
sent to the APIs and routing matrices will be chopped up into submatrices
of size 15x100. This option is necessary for freemium licenses to avoid
hitting the rate limit of the APIs with status code 429. Deactivate this
option to increase speed of requests for paid plans.
}
\examples{
set_freemium(FALSE)
}
