% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/authentication.R
\name{set_key}
\alias{set_key}
\title{Set HERE Application Credentials}
\usage{
set_key(api_key)
}
\arguments{
\item{api_key}{character, the API key from a HERE project.}
}
\value{
None.
}
\description{
Provide an API Key for a HERE project of type 'REST'.
The key is set for the current R session and is used
to authenticate in the requests to the APIs.
}
\details{
No login yet? Get a free login and key here: \href{https://developer.here.com/}{klick}
}
\examples{
set_key("<YOUR API KEY>")
}
