% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/authentication.R
\name{set_proxy}
\alias{set_proxy}
\title{Proxy Configuration}
\usage{
set_proxy(proxy, proxyuserpwd)
}
\arguments{
\item{proxy}{character, the URL of the proxy ("https://your-proxy.net:port/").}

\item{proxyuserpwd}{character, user and password for the authentication ("user:pwd").}
}
\description{
If a proxy is needed, for example because the computer is behind a corporate proxy,
it can be set as follows: proxy = "http://your-proxy.net:port/" or
"https://your-proxy.net:port/" and "proxyuserpwd" = "user:pwd".
}
\examples{
set_proxy("https://your-proxy.net:port/", "user:pwd")
}
