% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{set_rate_limit}
\alias{set_rate_limit}
\title{Limit requests to the APIs}
\usage{
set_rate_limit(ans = TRUE)
}
\arguments{
\item{ans}{boolean, use limits or not (default = \code{TRUE})?}
}
\value{
None.
}
\description{
If set to \code{TRUE} the hereR package limits the requests per second (RPS)
sent to the APIs. This option is necessary for freemium licenses to avoid
hitting the rate limit of the APIs with status code 429. Deactivate this
option to increase speed of requests for paid plans.
}
\examples{
set_rate_limit(FALSE)
}
