% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{set_verbose}
\alias{set_verbose}
\title{Verbose API usage of hereR}
\usage{
set_verbose(ans = FALSE)
}
\arguments{
\item{ans}{boolean, verbose or not (default = \code{FALSE})?}
}
\value{
None.
}
\description{
If set to \code{TRUE} the hereR package is messaging information about
the amount of requests sent to the APIs and data size received.
}
\examples{
set_verbose(TRUE)
}
