% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traffic.R
\name{traffic}
\alias{traffic}
\title{HERE Traffic API: Flow and Incidents}
\usage{
traffic(aoi, product = "flow", from_dt = NULL, to_dt = NULL,
  local_time = FALSE, url_only = FALSE)
}
\arguments{
\item{aoi}{\code{sf} object, Areas of Interest (POIs) of geometry type \code{POLYGON}.}

\item{product}{character, traffic product of the 'Traffic API'. Supported products: \code{"flow"} and \code{"incidents"}.}

\item{from_dt}{datetime, timestamp of type \code{POSIXct}, \code{POSIXt} for the earliest incident.}

\item{to_dt}{datetime, timestamp of type \code{POSIXct}, \code{POSIXt} for the latest incident.}

\item{local_time}{boolean, should time values in the response for traffic incidents be in the local time of the incident or in UTC (\code{default = FALSE})?}

\item{url_only}{boolean, only return the generated URLs (\code{default = FALSE})?}
}
\value{
An \code{sf} object containing the requested traffic information.
}
\description{
Traffic flow and incident information based on the 'Traffic API'.
The traffic flow data contains speed (\code{"SP"}) and congestion (jam factor: \code{"JF"}) information.
Traffic incidents contain information about location, time, duration, severity, description and other details.
}
\examples{
\donttest{
traffic(aoi = aoi[aoi$code == "LI", ], product = "flow")
}
}
\references{
\href{https://developer.here.com/api-explorer/rest/traffic}{HERE Traffic API}
}
