% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defunct.R
\name{unset_auth-defunct}
\alias{unset_auth-defunct}
\title{Remove Application Credentials}
\usage{
unset_auth()
}
\value{
None.
}
\description{
Remove previously set HERE application credentials from the current R session.
}
\seealso{
\code{\link{hereR-defunct}}
}
\keyword{internal}
