% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/authentication.R
\name{unset_auth}
\alias{unset_auth}
\title{Remove HERE Application Credentials}
\usage{
unset_auth()
}
\value{
None.
}
\description{
Remove previously set HERE application credentials from the current R session.
}
\examples{
unset_auth()
}
