% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/authentication.R
\name{unset_key}
\alias{unset_key}
\title{Remove HERE Application Credentials}
\usage{
unset_key()
}
\value{
None.
}
\description{
Remove previously set HERE API key from the current R session.
}
\examples{
unset_key()
}
