% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defunct.R
\name{unset_proxy-defunct}
\alias{unset_proxy-defunct}
\title{Remove Proxy Configuration}
\usage{
unset_proxy()
}
\value{
None.
}
\description{
Remove a previously set proxy configuration from the current R session.
}
\seealso{
\code{\link{hereR-defunct}}
}
\keyword{internal}
