% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defunct.R
\name{unset_proxy}
\alias{unset_proxy}
\title{#' @rdname hereR-defunct}
\usage{
unset_proxy()
}
\description{
#' @rdname hereR-defunct
}
\section{\code{unset_proxy}}{

For \code{unset_proxy}, configure a global proxy for R in '~/.Renviron' instead.
}

