% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{aoi}
\alias{aoi}
\title{Example AOIs}
\format{An object of class \code{"sf", "data.frame"}.}
\source{
Made with Natural Earth. Free vector and raster map data \href{https://raw.githubusercontent.com/datasets/geo-countries/master/data/countries.geojson}{@naturalearthdata.com}
}
\usage{
data(aoi)
}
\description{
Some example Areas of Interest (AOIs):
The boundary polygons of Switzerland and Liechtenstein.
}
\examples{
data(aoi)
}
\keyword{datasets}
