% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defunct.R
\name{hereR-defunct}
\alias{hereR-defunct}
\alias{set_auth}
\alias{unset_auth}
\title{Defunct functions in package \pkg{hereR}.}
\usage{
set_auth(app_id, app_code)

unset_auth()
}
\description{
The functions listed below are defunct.
  When possible, alternative functions with similar
  functionality are also mentioned. Help pages for defunct functions are
  available at \code{help("-defunct")}.
}
\section{\code{set_auth}}{

For \code{set_auth}, use \code{\link{set_key}}.
}

\section{\code{unset_auth}}{

For \code{unset_auth}, use \code{\link{unset_key}}.
}

\keyword{internal}
