% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{hereR-deprecated}
\alias{hereR-deprecated}
\title{Deprecated functions in package \pkg{hereR}.}
\description{
The functions listed below are deprecated and will be defunct in
  the near future. When possible, alternative functions with similar
  functionality are also mentioned. Help pages for deprecated functions are
  available at \code{help("-deprecated")}.
}
\keyword{internal}
