% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hereR-package.R
\docType{package}
\name{hereR-package}
\alias{hereR}
\alias{hereR-package}
\title{hereR: 'sf'-Based Interface to the 'HERE' REST APIs}
\description{
Interface to the 'HERE' REST APIs <https://developer.here.com/develop/rest-apis>: (1) geocode and autosuggest addresses or reverse geocode POIs using the 'Geocoder' API; (2) route directions, travel distance or time matrices and isolines using the 'Routing', 'Matrix Routing' and 'Isoline Routing' APIs; (3) request real-time traffic flow and incident information from the 'Traffic' API; (4) find request public transport connections and nearby stations from the 'Public Transit' API; (5) request intermodal routes using the 'Intermodal Routing' API; (6) get weather forecasts, reports on current weather conditions, astronomical information and alerts at a specific location from the 'Destination Weather' API. Locations, routes and isolines are returned as 'sf' objects.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://munterfi.github.io/hereR/}
  \item \url{https://github.com/munterfi/hereR/}
  \item Report bugs at \url{https://github.com/munterfi/hereR/issues/}
}

}
\author{
\strong{Maintainer}: Merlin Unterfinger \email{info@munterfinger.ch} (\href{https://orcid.org/0000-0003-2020-2366}{ORCID})

Other contributors:
\itemize{
  \item Daniel Possenriede (\href{https://orcid.org/0000-0002-6738-9845}{ORCID}) [contributor]
}

}
\keyword{internal}
