% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isoline.R
\name{isoline}
\alias{isoline}
\title{HERE Routing API: Isoline}
\usage{
isoline(points, range = seq(100, 1000, 100), rangetype = "distance",
  type = "fastest", mode = "car", traffic = FALSE,
  departure = NULL, start = TRUE, aggregate = TRUE)
}
\arguments{
\item{aggregate}{}
}
\description{
Calcuates a isoline (POLYGON) that connects the end points of all routes
leaving from one defined center with either a specified length or a
specified travel time.
}
\references{
\href{https://developer.here.com/documentation/routing/topics/resource-calculate-isoline.html}{HERE Routing API: Calculate Isoline}
}
