% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{poi}
\alias{poi}
\title{Example POIs}
\format{An object of class \code{"sf", "data.frame"}.}
\source{
Made with Natural Earth. Free vector and raster map data \href{https://raw.githubusercontent.com/nvkelso/natural-earth-vector/master/geojson/ne_10m_populated_places_simple.geojson}{@naturalearthdata.com}
}
\usage{
data(poi)
}
\description{
Some example Points of Interest (POIs):
Cities in Switzerland and Liechtenstein with more than 100'000 inhabitants.
}
\examples{
data(poi)
}
\keyword{datasets}
