% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/authentication.R
\name{unset_proxy}
\alias{unset_proxy}
\title{Remove Proxy Configuration}
\usage{
unset_proxy()
}
\value{
None.
}
\description{
Remove a previously set proxy configuration from the current R session.
}
\examples{
unset_proxy()
}
