% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weather.R
\name{weather}
\alias{weather}
\title{HERE Destination Weather API: Observation, Forecast or Astronomy}
\usage{
weather(poi, product = "observation", url_only = FALSE)
}
\arguments{
\item{product}{}
}
\value{
An sf object, containing the coordinates of the geocoded addresses.
}
\description{
Current weather conditions at Points of Interest (location or address).
}
\references{
\href{https://developer.here.com/documentation/weather/topics/example-weather-observation.html}{HERE Destination Weather API: Observation }
}
